finally the first Unit Test worked, for my PersonAddressController - full integration test, fixed the dbSeeder for the COuntrys, to get all the language codes and translations for all the countries in ISO code format, that google phoneLib is using, and some Mapper issues, finally have started to implement the Actions for WorkExp - not jet tested..

+++ backend/src/main/java/ibee/webapp/todo_app/core/repository/person/PersonSpecification.java
@@ -14,7 +14,7 @@ public class PersonSpecification {
             String firstName, 
             String lastName, 
             java.time.LocalDate birthDate, 
-            Short socialRecordNumber) {
+            Long socialRecordNumber) {
         
         return (root, query, cb) -> {
             List<Predicate> predicates = new ArrayList<>();
